Create day-9.1 (div tag + style)
Divine Separation
The <div> tag is used to divide an HTML page into sections of content. It is usually used as a container for other HTML tags. It is one of the most widely used tags. 

The block-level elements start on a new line.
If no width is set, it takes the full width available.
Note: Tags are also called Elements.
The div tag is a block-level element.
The style attribute can be added to a div tag to style it.